Add REST node parsing to adaptive-ui-designer-figma
