+++ packages/adaptive-ui-designer-core/src/index.ts
@@ -8,9 +8,12 @@ export {
     AdditionalDataKeys,
     DesignTokenValue,
     AppliedDesignToken,
-    AppliedStyleValue
+    AppliedStyleValue,
+    AdditionalData,
+    AppliedDesignTokens,
+    DesignTokenValues
 } from "./model.js";
-export { mapReplacer, mapReviver } from "./serialization.js";
-export { State, StatesState, PluginNode, focusIndicatorNodeName,  } from "./node.js";
+export { mapReplacer, mapReviver, deserializeMap } from "./serialization.js";
+export { State, StatesState, PluginNode, focusIndicatorNodeName, } from "./node.js";
 export { DesignTokenDefinition, DesignTokenRegistry, FormControlId } from "./registry/design-token-registry.js";
 export { nameToTitle, registerAppliableTokens, registerTokens } from "./registry/recipes.js";
