+++ packages/adaptive-ui-designer-core/tsconfig.json
@@ -9,5 +9,4 @@
         "outDir": "dist/esm",
         "rootDir": "src"
     },
-    "include": ["src"]
 }
