+++ packages/adaptive-ui-designer-figma/src/index.ts
@@ -1,0 +1,1 @@
+export { Anatomy, Condition, BooleanCondition, StringCondition, StyleRule } from "./anatomy.js";
