+++ packages/adaptive-ui-figma-designer/src/figma/tsconfig.json
@@ -7,9 +7,6 @@
         "lib": ["ES2018"],
         "strict": true,
         "skipLibCheck": true,
-        "typeRoots": [
-            "../../node_modules/@figma",
-            "../../node_modules/@types"
-        ]
+        "types": ["@figma/plugin-typings"],
     }
 }
