+++ tsconfig.json
@@ -11,7 +11,6 @@
         "resolveJsonModule": true,
         "sourceMap": true,
         "strictNullChecks": true,
-        "target": "ES2020",
-        "types": ["node"],
+        "target": "ES2020"
     }
 }
